♻️ VVQM9Y task: split workflow validation
